% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching_framework.R
\name{cached_compress}
\alias{cached_compress}
\title{Cached compression function}
\usage{
cached_compress(x, type = c("gzip", "bzip2", "xz"), ...)
}
\arguments{
\item{x}{A raw vector to compress}

\item{type}{Compression type}

\item{...}{Additional arguments passed to memCompress()}
}
\value{
Compressed raw vector
}
\description{
Cached wrapper for memCompress() function. Uses R.cache to store
results for repeated compressions with the same data and parameters.
}
