% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching_framework.R
\name{cached_spectrum}
\alias{cached_spectrum}
\title{Cached spectrum function}
\usage{
cached_spectrum(x, ...)
}
\arguments{
\item{x}{A univariate or multivariate time series}

\item{...}{Additional arguments passed to spectrum()}
}
\value{
An object of class "spec"
}
\description{
Cached wrapper for stats::spectrum() function. Uses R.cache to store
results for repeated calls with the same data and parameters.
}
