% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_reports.R
\name{create_visualizations}
\alias{create_visualizations}
\title{Create visualizations for discriminant analysis}
\usage{
create_visualizations(summary_data)
}
\arguments{
\item{summary_data}{Summary statistics data frame}
}
\value{
List of ggplot objects
}
\description{
Create visualizations for discriminant analysis
}
