% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package initialization function}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The library path where the package is installed}

\item{pkgname}{The name of the package ("qiprng")}
}
\description{
Called when the package is loaded. Initializes the package environment,
sets up the cryptographic backend (libsodium), and registers cleanup handlers.
}
\keyword{internal}
