% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_reports.R
\name{generate_detailed_report}
\alias{generate_detailed_report}
\title{Generate detailed report for top and bottom performers}
\usage{
generate_detailed_report(results, summary_data, n_top = 5, n_bottom = 5)
}
\arguments{
\item{results}{List of test results}

\item{summary_data}{Summary statistics data frame}

\item{n_top}{Number of top performers to report}

\item{n_bottom}{Number of bottom performers to report}
}
\value{
Character string with detailed report
}
\description{
Generate detailed report for top and bottom performers
}
