% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excellent_discriminants.R
\name{generate_excellent_random}
\alias{generate_excellent_random}
\title{Generate random numbers using an excellent discriminant}
\usage{
generate_excellent_random(
  n = 10000,
  discriminant_index = 1,
  results_file = "discriminant_analysis_results/raw_results.rds",
  precision = 256
)
}
\arguments{
\item{n}{Number of random numbers to generate}

\item{discriminant_index}{Index of discriminant to use (1 = best, 2 = second best, etc.)}

\item{results_file}{Path to the analysis results RDS file}

\item{precision}{MPFR precision (default 256)}
}
\value{
Vector of random numbers
}
\description{
Generate random numbers using an excellent discriminant
}
