% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_reports.R
\name{generate_summary_stats}
\alias{generate_summary_stats}
\title{Generate summary statistics from test results}
\usage{
generate_summary_stats(results)
}
\arguments{
\item{results}{List of test results from run_discriminant_analysis}
}
\value{
Data frame with summary statistics
}
\description{
Generate summary statistics from test results
}
