% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_tests.R
\name{generate_with_discriminant}
\alias{generate_with_discriminant}
\title{Generate random numbers using specific discriminant parameters}
\usage{
generate_with_discriminant(a, b, c, n = 50000, precision = 256)
}
\arguments{
\item{a}{Parameter a}

\item{b}{Parameter b}

\item{c}{Parameter c}

\item{n}{Number of samples to generate}

\item{precision}{MPFR precision (default 256)}
}
\value{
Vector of random numbers
}
\description{
Generate random numbers using specific discriminant parameters
}
