% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching_framework.R
\name{is_cache_enabled}
\alias{is_cache_enabled}
\title{Check if caching is enabled}
\usage{
is_cache_enabled()
}
\value{
Logical indicating if caching is enabled
}
\description{
Check if caching is enabled
}
