% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_suite.R
\name{print_validation_summary}
\alias{print_validation_summary}
\title{Print validation summary}
\usage{
print_validation_summary(report)
}
\arguments{
\item{report}{Validation report}
}
\description{
Print validation summary
}
