% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prng_interface.R, R/qiprng-package.R
\docType{package}
\name{qiprng}
\alias{qiprng}
\title{Enhanced Quadratic Irrational PRNG}
\description{
A package implementing a high-precision quadratic irrational PRNG
with optional cryptographic mixing, hardware RNG, automatic reseeding,
multiple distributions, etc.

Provides a configurable pseudo-random number generator based on
quadratic irrationals with optional cryptographic mixing.
}
\details{
The QIPRNG package implements a high-precision pseudo-random number
generator based on quadratic irrational numbers. It supports various
output distributions including uniform, normal, and exponential.
Optional cryptographic mixing can be applied for enhanced security.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/biostochastics/qiprng/issues}
}


Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/biostochastics/qiprng/issues}
}

}
\author{
\strong{Maintainer}: Sergey Kornilov \email{sergey.kornilov@biostochastics.com}


Sergey Kornilov
}
\keyword{internal}
