% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_reports.R
\name{save_analysis_results}
\alias{save_analysis_results}
\title{Save all reports and visualizations}
\usage{
save_analysis_results(results, output_dir = "discriminant_analysis_results")
}
\arguments{
\item{results}{List of test results}

\item{output_dir}{Directory to save outputs}
}
\description{
Save all reports and visualizations
}
