% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_testing.R
\name{summarize_test_results}
\alias{summarize_test_results}
\title{Generate a summary of test results}
\usage{
summarize_test_results(suite)
}
\arguments{
\item{suite}{The test suite with results}
}
\value{
The test suite with summary added
}
\description{
Generate a summary of test results
}
\keyword{internal}
