% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_tests.R
\name{test_autocorrelation}
\alias{test_autocorrelation}
\title{Test for autocorrelation}
\usage{
test_autocorrelation(samples, max_lag = 50)
}
\arguments{
\item{samples}{Vector of random samples}

\item{max_lag}{Maximum lag to test}
}
\value{
List with test results
}
\description{
Test for autocorrelation
}
