% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_tests.R
\name{test_independence}
\alias{test_independence}
\title{Test independence using runs test}
\usage{
test_independence(samples)
}
\arguments{
\item{samples}{Vector of random samples}
}
\value{
List with test results
}
\description{
Test independence using runs test
}
