% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_tests.R
\name{test_moments}
\alias{test_moments}
\title{Test moments (mean, variance, skewness, kurtosis)}
\usage{
test_moments(samples)
}
\arguments{
\item{samples}{Vector of random samples}
}
\value{
List with test results
}
\description{
Test moments (mean, variance, skewness, kurtosis)
}
