% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_tests.R
\name{test_poker}
\alias{test_poker}
\title{Poker test for randomness}
\usage{
test_poker(samples, m = 5)
}
\arguments{
\item{samples}{Vector of random samples}

\item{m}{Length of each hand (default 5)}
}
\value{
List with poker test results
}
\description{
Poker test for randomness
}
