% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_tests.R
\name{test_serial_correlation}
\alias{test_serial_correlation}
\title{Serial correlation test at multiple lags}
\usage{
test_serial_correlation(samples, max_lag = 50)
}
\arguments{
\item{samples}{Vector of random samples}

\item{max_lag}{Maximum lag to test}
}
\value{
List with serial correlation test results
}
\description{
Serial correlation test at multiple lags
}
