% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_tests.R
\name{test_uniformity}
\alias{test_uniformity}
\title{Test uniformity using Kolmogorov-Smirnov test}
\usage{
test_uniformity(samples, bins = 20)
}
\arguments{
\item{samples}{Vector of random samples}
}
\value{
List with test results
}
\description{
Test uniformity using Kolmogorov-Smirnov test
}
