% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_suite.R
\name{validate_basic_tests}
\alias{validate_basic_tests}
\title{Validate basic tests}
\usage{
validate_basic_tests(config, verbose = TRUE)
}
\arguments{
\item{config}{Validation configuration}

\item{verbose}{Print progress}
}
\value{
Validation results for basic tests
}
\description{
Validate basic tests
}
