% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_suite.R
\name{validate_binary_tests}
\alias{validate_binary_tests}
\title{Validate binary tests}
\usage{
validate_binary_tests(config, verbose = TRUE)
}
\arguments{
\item{config}{Validation configuration}

\item{verbose}{Print progress}
}
\value{
Validation results for binary tests
}
\description{
Validate binary tests
}
