% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_suite.R
\name{validate_external_tests}
\alias{validate_external_tests}
\title{Validate external tests}
\usage{
validate_external_tests(config, verbose = TRUE)
}
\arguments{
\item{config}{Validation configuration}

\item{verbose}{Print progress}
}
\value{
Validation results for external tests
}
\description{
Validate external tests
}
