% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_suite.R
\name{validate_multidim_tests}
\alias{validate_multidim_tests}
\title{Validate multidimensional tests}
\usage{
validate_multidim_tests(config, verbose = TRUE)
}
\arguments{
\item{config}{Validation configuration}

\item{verbose}{Print progress}
}
\value{
Validation results for multidimensional tests
}
\description{
Validate multidimensional tests
}
