% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_suite.R
\name{validate_performance}
\alias{validate_performance}
\title{Validate performance}
\usage{
validate_performance(config, verbose = TRUE)
}
\arguments{
\item{config}{Validation configuration}

\item{verbose}{Print progress}
}
\value{
Performance validation results
}
\description{
Validate performance
}
