% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_suite.R
\name{withTimeout}
\alias{withTimeout}
\title{Timeout helper function}
\usage{
withTimeout(expr, timeout)
}
\arguments{
\item{expr}{Expression to evaluate}

\item{timeout}{Timeout in seconds}
}
\description{
Timeout helper function
}
