% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_suite.R
\name{modifyList}
\alias{modifyList}
\title{Helper function to modify lists}
\usage{
modifyList(x, val)
}
\arguments{
\item{x}{Base list}

\item{val}{List of modifications}
}
\value{
Modified list
}
\description{
Helper function to modify lists
}
