% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excellent_discriminants.R
\name{print_excellent_summary}
\alias{print_excellent_summary}
\title{Print summary of excellent discriminants analysis}
\usage{
print_excellent_summary(
  results_file = "discriminant_analysis_results/raw_results.rds"
)
}
\arguments{
\item{results_file}{Path to the analysis results RDS file}
}
\description{
Print summary of excellent discriminants analysis
}
