% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_tests.R
\name{run_advanced_tests}
\alias{run_advanced_tests}
\title{Comprehensive advanced test suite}
\usage{
run_advanced_tests(samples)
}
\arguments{
\item{samples}{Vector of random samples}
}
\value{
List with all advanced test results
}
\description{
Comprehensive advanced test suite
}
