% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant_tests.R
\name{test_discriminant}
\alias{test_discriminant}
\title{Comprehensive test suite for a single discriminant}
\usage{
test_discriminant(a, b, c, discriminant, n = 50000)
}
\arguments{
\item{a}{Parameter a}

\item{b}{Parameter b}

\item{c}{Parameter c}

\item{discriminant}{Discriminant value}

\item{n}{Number of samples to generate}
}
\value{
List with all test results
}
\description{
Comprehensive test suite for a single discriminant
}
