% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_tests.R
\name{test_entropy}
\alias{test_entropy}
\title{Shannon entropy test}
\usage{
test_entropy(samples, bins = 256)
}
\arguments{
\item{samples}{Vector of random samples}

\item{bins}{Number of bins for discretization}
}
\value{
List with entropy test results
}
\description{
Shannon entropy test
}
