% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prng_interface.R
\name{validate_config}
\alias{validate_config}
\title{Validate PRNG configuration}
\usage{
validate_config(config)
}
\arguments{
\item{config}{List of configuration parameters}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate PRNG configuration
}
\keyword{internal}
